+++ test/repository/remote/test_orders.py
@@ -3,58 +3,9 @@ from typing import List
 import pendulum
 import pytest
 
-from conf.consts import TimeInForce, OrderType, Side, CryptoAsset, OrderStatus
+from conf.consts import OrderType, Side
 from repository import TradingPair
 from repository.db import Fill, Order
-
-
-@pytest.fixture
-def api_response() -> dict:
-    return {
-        "symbol": "BNBBUSD",
-        "orderId": 271468,
-        "orderListId": -1,
-        "clientOrderId": "AS8MQlzMS8vEGwobLGWpN5",
-        "transactTime": 1665544366616,
-        "price": "0.00000000",
-        "origQty": "10.00000000",
-        "executedQty": "10.00000000",
-        "cummulativeQuoteQty": "2714.61500000",
-        "status": "FILLED",
-        "timeInForce": "GTC",
-        "type": "MARKET",
-        "side": "SELL",
-        "fills": [
-            {
-                "price": "271.70000000",
-                "qty": "3.50000000",
-                "commission": "0.00000000",
-                "commissionAsset": "BUSD",
-                "tradeId": 11243,
-            },
-            {
-                "price": "271.50000000",
-                "qty": "3.14000000",
-                "commission": "0.00000000",
-                "commissionAsset": "BUSD",
-                "tradeId": 11244,
-            },
-            {
-                "price": "271.20000000",
-                "qty": "2.59000000",
-                "commission": "0.00000000",
-                "commissionAsset": "BUSD",
-                "tradeId": 11245,
-            },
-            {
-                "price": "271.10000000",
-                "qty": "0.77000000",
-                "commission": "0.00000000",
-                "commissionAsset": "BUSD",
-                "tradeId": 11246,
-            },
-        ],
-    }
 
 
 @pytest.fixture
@@ -91,9 +42,10 @@ def liquid_symbol() -> TradingPair:
     return TradingPair.structure("BNBUSDT")
 
 
-def test_place_test_order(api_client, liquid_symbol):
+def test_place_test_order(api_client, liquid_symbol, env_state_id):
     assert (
         api_client.place_order(
+            env_state_id=env_state_id,
             pair=liquid_symbol,
             side=Side.SELL,
             type=OrderType.MARKET,
@@ -103,10 +55,11 @@ def test_place_test_order(api_client, liquid_symbol):
     )
 
 
-def test_place_order(api_client, struct_order, liquid_symbol):
+def test_place_order(api_client, struct_order, liquid_symbol, env_state_id):
     # Test fixed values
     now = pendulum.now()
     order = api_client.place_order(
+        env_state_id=env_state_id,
         pair=liquid_symbol,
         side=struct_order.side,
         type=struct_order.type,
@@ -131,5 +84,5 @@ def test_structure_fills(uns_fills, struct_fills):
     assert Fill.structure(uns_fills) == struct_fills
 
 
-def test_structure_order(api_response, struct_order):
-    assert Order.structure(api_response) == struct_order
+def test_structure_order(api_response, struct_order, env_state_id):
+    assert Order.structure({**{"env_state_id": env_state_id}, **api_response}) == struct_order
